Refactor security event helpers
